% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageassn-package.r
\docType{package}
\name{ageassn}
\alias{ageassn}
\alias{ageassn-package}
\alias{ageassn-package}
\title{ageassn: Running Analyses for the BrCa Age Associated Biomarker Project}
\description{
Various data setup, munging, and processing functions used to
obtain results for METABRIC and TCGA datasets of the BrCa Age Associated
Biomarker project. Result tables, scatterplots, volcano plots, and manhattan
plots comprise some of the outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BCCRCMO/ageassn}
  \item Report bugs at \url{https://github.com/BCCRCMO/ageassn/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

}
\keyword{internal}
