% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_boxplot}
\alias{boot_boxplot}
\title{Boxplot of proportion of significant results per group}
\usage{
boot_boxplot(data, title = NULL)
}
\arguments{
\item{data}{data object from \code{boot_lm}}

\item{title}{plot title}
}
\description{
Boxplot for each group, showing distribution of bootstrap proportions
}
