% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_generate}
\alias{boot_generate}
\title{Generate bootstrapped data}
\usage{
boot_generate(data, group, seed = 1, reps = 20, nb = NULL)
}
\arguments{
\item{data}{data object with model variables of interest}

\item{group}{a factor that splits the samples in \code{data} into different groups}

\item{seed}{random seed for reproducibility}

\item{reps}{number of bootstrap replicates}

\item{nb}{an integer stating the bootstrap sample size for each group. If
\code{NULL} (default), the bootstrap sample size for each group is taken from
the original group-specific sample sizes.}
}
\value{
A list with length equal to the number of levels in \code{group}. If \code{nb}
is not \code{NULL}, each element itself is a list of \code{reps} data frames with
\code{nb} rows. If \code{nb} is \code{NULL}, each element is a list of length 1 with a
data frame whose number of rows is equal to its original group-size.
}
\description{
Generate bootstrapped data within each group for a specified number of
bootstrap samples and replicates.
}
\details{
The \code{data} is first split by each level in the \code{group} factor. Then, \code{reps}
replicates of bootstrap samples of size \code{nb} are generated for each data
subset. If no \code{nb} is specified, the original group-specific sizes are used
and no bootstrapping is performed.
}
\examples{
str(boot_generate(iris, iris$Species, reps = 2, nb = 100), max.level = 2)
str(boot_generate(iris, iris$Species), max.level = 2)
}
\author{
Derek Chiu
}
