% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_lm}
\alias{boot_lm}
\title{Linear models on bootstrapped data}
\usage{
boot_lm(data, x, y, group, probes, genes, aridxs, age = 60, fdr = 0.01,
  fc = 1.25, seed = 1, reps = 20, nb = NULL, filename = NULL)
}
\arguments{
\item{data}{data object with model variables of interest}

\item{x}{charater string of the explanatory variable in \code{data}}

\item{y}{data object from which the variables named in \code{probes} are found}

\item{group}{a factor that splits the samples in \code{data} into different groups}

\item{probes}{character vector of probe names to model}

\item{genes}{character vector of gene names represented by the \code{probes}}

\item{aridxs}{age-related probes of interest}

\item{age}{linear model performed on the datasets \code{x <= age}, \code{x > age}, and
the full model}

\item{fdr}{FDR threshold for BH-adjusted p-values statistical significance}

\item{fc}{fold change threshold for biological significance}

\item{seed}{random seed for reproducibility}

\item{reps}{number of bootstrap replicates}

\item{nb}{an integer stating the bootstrap sample size for each group. If
\code{NULL} (default), the bootstrap sample size for each group is taken from
the original group-specific sample sizes.}

\item{filename}{save result object to file}
}
\value{
A list of length \code{reps}, each of which is a matrix with
\code{nlevels(group)} rows and two columns, showing for each group, the number
of response variables in \code{y} that showed a biologically relevant and
statistically significant result.
}
\description{
Run linear regression models on bootstrapped data
}
\details{
This function is a wrapper around \code{boot_generate}, \code{boot_stats}, and
\code{boot_counts}.
}
\author{
Derek Chiu
}
