% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_stats}
\alias{boot_stats}
\title{Calculate statistics from bootstrapped data}
\usage{
boot_stats(boot_data, x, y, probes, genes, aridxs, age = 60, fdr = 0.01,
  fc = 1.25)
}
\arguments{
\item{boot_data}{bootstrapped data from \code{boot_generate}}

\item{x}{charater string of the explanatory variable in \code{data}}

\item{y}{data object from which the variables named in \code{probes} are found}

\item{probes}{character vector of probe names to model}

\item{genes}{character vector of gene names represented by the \code{probes}}

\item{aridxs}{age-related probes of interest}

\item{age}{linear model performed on the datasets \code{x <= age}, \code{x > age}, and
the full model}

\item{fdr}{FDR threshold for BH-adjusted p-values statistical significance}

\item{fc}{fold change threshold for biological significance}
}
\value{
A nested list of length equal to the number of replicates specified
in the generation of \code{boot_data}. Each element is itself a list of data
frames, one for each of the groups in \code{boot_data}. In essence, the data
structure is a transpose of that found in \code{boot_data}. The number of rows
in each of the statistical result data frames is equal to the number of
\code{probes} (and \code{genes}).
}
\description{
Calculate linear model statistics like p-values and slopes from bootstrapped
data generated by \code{boot_generate}.
}
\details{
The age-associated linear model for METABRIC expression is run on each of the
bootstrapped data objects from \code{boot_generate}. The slopes, adjusted
p-values, and other statistics for the less than or equal to \code{age}, greater
than \code{age}, and all cases groups are returned.

The results are placed in a nested list, since we want them separated by
group and replicate. Each of the results is a data frame with 28 variables:
\itemize{
\item \code{PROBE}: probe id
\item \code{GENE}: gene name
\item \code{LE_slope}: estimated slope for less than or equal to \code{age}
\item \code{GT_slope}: estimated slope for greater than \code{age}
\item \code{All_slope}: estimated slope for full data
\item \code{LE_pval}: p-value for less than or equal to \code{age}
\item \code{GT_pval}: p-value for less greater than \code{age}
\item \code{All_pval}: p-value for full data
\item \code{AgeDependentp}: logical variable for age dependent
\item \code{LE_pval_BHadj}: BH-adjusted p-value for less than or equal to \code{age}
\item \code{GT_pval_BHadj}: BH-adjusted p-value for greater than \code{age}
\item \code{All_pval_BHadj}: BH-adjusted p-value for full data
\item \code{BHadj_and_AgeDependentp}: logical variable for statistical significance
at BH-adjusted p-values and age dependence
\item \code{BHadj_signifp}: logical variable for statistical significance at
BH-adjusted p-values
\item \code{LE_log2FC}: log2 fold change for less than or equal \code{age}
\item \code{GT_log2FC}: log2 fold change for greater than \code{age}
\item \code{All_log2FC}: log2 fold change for full data
\item \code{Best_log2FC}: the log2 fold change with the strongest significance
\item \code{Best_BHadj_pval}: the BH-adjusted p-value with the strongest significance
\item \code{Abs_Best_log2FC}: absolute value of \code{Best_log2FC}
\item \code{Abs_FoldChange}: \code{2 ^ Abs_Best_log2FC}
\item \code{FoldChange_Direction}: "UP" if \code{Best_log2FC > 0}, "DOWN" otherwise
\item \code{arGeneSetidxp}: logical variable for probes found in \code{aridxs}
\item \code{arGeneSet_LE_BHadj_pval}: BH-adjusted p-value for less than or equal to
\code{age} if probe is in \code{aridxs}, \code{NA} otherwise
\item \code{arGeneSet_GT_BHadj_pval}: BH-adjusted p-value for greater than \code{age} if
probe is in \code{aridxs}, \code{NA} otherwise
\item \code{arGeneSet_All_BHadj_pval}: BH-adjusted p-value for full data if probe is
in \code{aridxs}, \code{NA} otherwise
\item \code{arGeneSet_BHadj_and_AgeDependentp}: logical variable for statistical
significance at BH-adjusted p-values and age dependence if found in \code{aridxs},
\code{NA} otherwise
\item \code{arGeneSet_PBHadj}: the smallest of \code{arGeneSet_LE_BHadj_pval},
\code{arGeneSet_GT_BHadj_pval}, and \code{arGeneSet_All_BHadj_pval}
}
}
\author{
Derek Chiu
}
