% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_summarize}
\alias{boot_summarize}
\title{Summarize bootstrap analysis}
\usage{
boot_summarize(data, fun = c("mean", "median"))
}
\arguments{
\item{data}{data object from \code{boot_lm}}

\item{fun}{summarize bootstrap statistics using the "mean" or "median"}
}
\description{
The trend and no_trend counts across replicates using different averaging
statistics
}
