% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_top}
\alias{boot_top}
\title{Top probes from bootstrap analysis}
\usage{
boot_top(stat_data, group, freqs)
}
\arguments{
\item{stat_data}{statistics data from \code{boot_stats}}

\item{group}{a factor that splits the samples in \code{data} into different groups}

\item{freqs}{thresholds of frequencies for top probes}
}
\description{
List of top probes chosen most frequently out of all bootstrap replicates.
}
\details{
To quantify what is meant by most frequent, the \code{freqs} vector serves as a
threshold. For example, if \code{freqs = c(10, 15, 19, 20)}, then we are
interested in the probes that were chosen in 10, 15, 19, 20 (out of 20)
replicates.
}
