% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ageassn.R
\name{comp_metabric_tcga}
\alias{comp_metabric_tcga}
\title{Compare age association across cancer datasets}
\usage{
comp_metabric_tcga(metabric, tcga, fdr, fc, organ = "Br")
}
\arguments{
\item{metabric}{METABRIC dataset object}

\item{tcga}{list of TCGA dataset objects. Must have specific naming syntax.}

\item{fdr}{false discovery rate}

\item{fc}{fold change}

\item{organ}{TCGA organ determines how \code{tcga} is subsetted for the
corresponding \code{fdr} and \code{fc} levels.}
}
\value{
An object of class "CrossTable" comparing the age association across
different cancer datasets with a contingency table and different hypothesis
tests of independence.
}
\description{
Compare age associated status (presence/absence) between TCGA and METABRIC
datasets.
}
