% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ageassn.R
\name{comp_pair_tcga}
\alias{comp_pair_tcga}
\title{Compare pairs of TCGA data on age association}
\usage{
comp_pair_tcga(x, y, x_organ, y_organ, fdr, fc, N = 18950)
}
\arguments{
\item{x}{first list of TCGA data}

\item{y}{second list of TCGA other}

\item{x_organ}{first TCGA organ}

\item{y_organ}{second TCGA organ}

\item{fdr}{false discovery rate}

\item{fc}{fold change}

\item{N}{total number of cases}
}
\description{
Compare pairs of TCGA data on age association
}
