% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{is_mean_enclosed}
\alias{is_mean_enclosed}
\title{Check mean enclosure}
\usage{
is_mean_enclosed(data, x, y, conf.level = 0.99)
}
\arguments{
\item{data}{data object}

\item{x}{age variable in \code{data}}

\item{y}{expression variable in \code{data}}

\item{conf.level}{confidence level}
}
\value{
Returns \code{TRUE} only if for every set of upper and lower pointwise
confidence limits, the mean expression is between these bounds, \code{FALSE} if
the mean expression falls outside of the confidence limits for at least one
point.
}
\description{
Is the mean expression fully enclosed by the loess confidence bands?
}
\examples{
is_mean_enclosed(mtcars, "wt", "qsec", conf.level = 0.99)
is_mean_enclosed(mtcars, "wt", "qsec", conf.level = 0.95)
}
\author{
Derek Chiu
}
