% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan Plot}
\usage{
manhattan_plot(x, CHR = "CHR", BP = "BP", PBHadj = "PBHadj",
  snp = "SNP", logp = TRUE, chrlabs = NULL, col = c("black", "red3",
  "green4", "blue", "cyan3", "magenta3", "orange", "gray40"),
  suggestiveline = 0.05, genomewideline = FALSE, highlight = FALSE,
  highlight.var = NULL, plotpointhiliteidxp = FALSE, hilitelbls = NULL,
  ...)
}
\arguments{
\item{x}{data object}

\item{CHR}{chromosome numbers}

\item{BP}{base pairs}

\item{PBHadj}{BH-adjusted p-values}

\item{snp}{SNP gene names}

\item{logp}{logical; if \code{TRUE} use -log10 p-values instead of raw}

\item{chrlabs}{labels for chromosomes}

\item{col}{colours to use for chromosomes; can be recycled}

\item{suggestiveline}{line for suggestive association}

\item{genomewideline}{line for genomewide association}

\item{highlight}{logical; if \code{TRUE}, points annotated by shape and label}

\item{highlight.var}{variable name to use highlighting}

\item{plotpointhiliteidxp}{whether to highlight points by shape}

\item{hilitelbls}{labels to use for highlighting}

\item{...}{additional parameters to \code{plot}}
}
\value{
A manhattan plot of chromosome number vs adjusted p-value
}
\description{
Manhattan Plot
}
\references{
https://github.com/stephenturner/qqman/blob/master/R/manhattan.R
}
\author{
Derek Chiu
}
