% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{max_limits}
\alias{max_limits}
\alias{max_xlim}
\alias{max_limits}
\alias{max_ylim}
\title{Maximum axes limits}
\usage{
max_xlim(.x, var = "Best_log2FC")

max_ylim(.x, var = "PBHadj_log")
}
\arguments{
\item{.x}{list of datasets}

\item{var}{variable name}
}
\description{
Calculate maximum x-axis/y-axis limits across list of datasets for
plotting equal axes across a subgroup.
}
