% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{MBid2Ctr}
\alias{MBid2Ctr}
\title{Take in any ID coding and return treatment centre}
\usage{
MBid2Ctr(x)
}
\arguments{
\item{x}{METABRIC ID}
}
\description{
Take in any ID coding and return treatment centre
}
