% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_ar}
\alias{metabric_ar}
\title{Add arGeneSet variables to aroutdf}
\usage{
metabric_ar(x, aridxs, annodf, age = 60, fdr = 0.01)
}
\arguments{
\item{x}{aroutdf object}

\item{aridxs}{ar gene set}

\item{annodf}{annotation data}

\item{age}{age threshold for calculating biological significance}

\item{fdr}{false discovery rate level}
}
\value{
extra variables added to \code{x}. Nothing is done if \code{x} had no
\code{BHadj_and_AgeDependentp} probes.
}
\description{
Add arGeneSet variables to aroutdf
}
