% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_aroutdf}
\alias{metabric_aroutdf}
\title{Make aroutdf output object}
\usage{
metabric_aroutdf(data, filename, annodf, ebdf, Dataset_r, age = 60,
  fdr = 0.01, fc = c(1.25, 2, 4), group = NULL)
}
\arguments{
\item{data}{clinical data}

\item{filename}{file name}

\item{annodf}{annotation data}

\item{ebdf}{ER binding genes}

\item{Dataset_r}{probe data}

\item{age}{age cutoff for analysis. Defaults to 60.}

\item{fdr}{false discovery rate. Defaults to 0.01.}

\item{fc}{fold change. Defaults to 1.25, 2, and 4 (3 sets of results).}

\item{group}{if \code{NULL}, results are for all cases. Otherwise, the variable
name for which to split \code{data} by.}
}
\value{
aroutdf object used for determining age-dependent trends
}
\description{
Make aroutdf output object
}
