% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_ebdf}
\alias{metabric_ebdf}
\title{Add additional ER binding gene lists to aroutdf}
\usage{
metabric_ebdf(x, ebdf_t1, ebdf_t2)
}
\arguments{
\item{x}{aroutdf object}

\item{ebdf_t1}{Tier 1 ER binding gene set from NKI}

\item{ebdf_t2}{Tier 2 ER binding gene set from NKI}
}
\value{
extra variables added to \code{x}
}
\description{
Add additional ER binding gene lists to aroutdf
}
