% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_evcp}
\alias{metabric_evcp}
\title{Add evcp variable}
\usage{
metabric_evcp(x, thresholds)
}
\arguments{
\item{x}{aroutdf object}

\item{thresholds}{a vector of length six, for three sets of logFC and
BH-adjusted p-value conditions. The 1st and 2nd element comprise the first
condition, the 3rd and 4th element comprise the second condition, the 5th
and 6th condition comprise the third condition.}
}
\description{
Set conditions for highligting METABRIC volcano plot points
}
