% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_fet}
\alias{metabric_fet}
\title{METABRIC Fisher Exact Tables}
\usage{
metabric_fet(data, probes, erbnms_t1, erbnms_t2)
}
\arguments{
\item{data}{probe data}

\item{probes}{vector of probes to keep matching TCGA Entrez gene IDs}

\item{erbnms_t1}{distinct Tier 1 ER binding gene names from NKI}

\item{erbnms_t2}{distinct Tier 2 ER binding gene names from NKI}
}
\value{
A contingency table of NKI ER binding gene status and age associated
status
}
\description{
METABRIC Fisher Exact Tables
}
