% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{metabric_signif}
\alias{metabric_signif}
\title{Set the biological significance variable in output data}
\usage{
metabric_signif(data, var, slopes = NULL, pvalues = NULL, fc = 1.25,
  fdr = 0.01)
}
\arguments{
\item{data}{aroutdf data}

\item{var}{variable name of biological significance}

\item{slopes}{slope variable names for Less than or Equal, Greater Than, and
All cases}

\item{pvalues}{adjusted p-value variable names for Less than or Equal,
Greater Than, and All cases}

\item{fc}{fold change}

\item{fdr}{false discovery rate}
}
\value{
modified data object
}
\description{
Modify the data object's biological significance variable based on
fold change and false discovery rate thresholds.
}
\details{
The default minimum fold change of biological significance is 1.25 and the
false discovery rate set to 0.01. One of three conditions based on threshold
of the explanatory variable (e.g. age cutoff) must be true for biological
significance: fold change greater than its biological significance constant
and adjusted p-value less than false discovery rate for either the lower
cutoff, upper cutoff, or the entire range of the explanatory variable.
}
