% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{metabric_volcano}
\alias{metabric_volcano}
\title{Make a volcano plot specifically for data structures in METABRIC studies}
\usage{
metabric_volcano(data, root.dir, plot.suffix, title, ...)
}
\arguments{
\item{data}{list of data objects for plotting}

\item{root.dir}{root directory to save plots}

\item{plot.suffix}{suffix at end of plot file name}

\item{title}{plot title}

\item{...}{named arguments to pass to \code{volcano_plot}}
}
\description{
Make a volcano plot specifically for data structures in METABRIC studies
}
