% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{refSeqRowMatches_annodf}
\alias{refSeqRowMatches_annodf}
\title{Row match with annodf}
\usage{
refSeqRowMatches_annodf(pattern, cols, annodf)
}
\arguments{
\item{pattern}{matching pattern}

\item{cols}{columns to match on}

\item{annodf}{annotation data}
}
\description{
Row match with annodf
}
