% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{regression_plot}
\alias{regression_plot}
\title{Fit a full linear model with subsetting at age of interest}
\usage{
regression_plot(data, x, y, age, fc = 1.25, fdr = 0.01, main = NULL,
  xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  xlim.pval = NULL, ylim.pval = NULL, annotate = NULL,
  xlim.trend = NULL, ylim.trend = NULL, conf.level = NULL,
  show.avg = FALSE, erbtxt = NULL)
}
\arguments{
\item{data}{data object}

\item{x}{age variable in \code{data}}

\item{y}{expression variable in \code{data}}

\item{age}{where to cut \code{x} for subsetted age analysis}

\item{fc}{fold change}

\item{fdr}{false discovery rate}

\item{main}{plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{xlim.pval}{horizontal limits for p-values}

\item{ylim.pval}{vertical limits for p-values}

\item{annotate}{vector of genes with an age-dependent trend that can be
annotated}

\item{xlim.trend}{horizontal limits for age-dependent trend annotation}

\item{ylim.trend}{vertical limits for age-dependent trend annotation}

\item{conf.level}{confidence level}

\item{show.avg}{show average?}

\item{erbtxt}{ER binding annotation text}
}
\value{
linear model fit
}
\description{
Fit a full linear model with subsetting at age of interest
}
\author{
Derek Chiu, Steven McKinney
}
