% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabric.R
\name{rzscore.strat.transform}
\alias{rzscore.strat.transform}
\title{Robust Z-Score stratified transformation}
\usage{
rzscore.strat.transform(datamat, stratavec)
}
\arguments{
\item{datamat}{data matrix}

\item{stratavec}{vector of strata generated by \code{MBid2Ctr}}
}
\description{
Robust Z-Score stratified transformation
}
