% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterfit_plot.R
\name{scatterfit_plot}
\alias{scatterfit_plot}
\title{Scatter-fit plot}
\usage{
scatterfit_plot(x, age = 60, title = NULL, caption = NULL)
}
\arguments{
\item{x}{data object}

\item{age}{age of interest}

\item{title}{plot title. Defaults to gene name and shows ER binding status if
present}

\item{caption}{plot caption}
}
\value{
A scatterplot with fit trend line added
}
\description{
Raw expression vs. age with smoother and annotations
}
\author{
Derek Chiu
}
