% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_file}
\alias{str_file}
\alias{str_decimal}
\alias{str_file}
\alias{str_sign}
\title{Convert strings for use in file names}
\usage{
str_decimal(string)

str_sign(string, type = c("short", "long"))
}
\arguments{
\item{string}{character string}

\item{type}{use "short" or "long" type of conversion}
}
\description{
Certain characters are not advisable for use in file names. \code{str_decimal}
converts decimal numbers and \code{str_sign} converts positive and negative signs.
}
\details{
Decimal numbers contain a period character ".", which is traditionally used
to separate the name and extension of a file. \code{str_decimal} replaces all
periods with "p" (for "period").

The plus sign "+", minus sign "-", and division sign "/" are all not
recommended for use in file names. There are two syntaxes used for
replacement. In the short version, they are replaced by "p", "n", and "_",
respectively. In the long version, they are replaced by "pos" (for
"positive"), "neg" (for "negative"), and "", respectively.
}
\examples{
str_decimal(1.23)
str_sign("ER+/HER2-")
str_sign("ER+/HER2-", type = "long")
}
