% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_aroutdf}
\alias{tcga_aroutdf}
\title{Create aroutdf object indicating p-values, estimates, age-dependent trend
indicators from the regression result}
\usage{
tcga_aroutdf(exandf, sedf, erbnms, root.dir, age = 60, fdr = 0.01,
  fc = 1.25, subset = FALSE, overwrite = FALSE, file.name = NULL,
  suffix = NULL)
}
\arguments{
\item{exandf}{expression data}

\item{sedf}{analytical data}

\item{erbnms}{ER binding names}

\item{root.dir}{root directory of aroutdf object}

\item{age}{age threshold for calculating biological significance}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}

\item{subset}{if \code{TRUE}, also save subsetted aroutdf results}

\item{overwrite}{if \code{TRUE}, save result even if it already exists}

\item{file.name}{core file name}

\item{suffix}{suffix for file name}
}
\value{
The object named aroutdf, with columns "Hugo_Symbol",
"Entrez_Gene_Id", and the statistical results
}
\description{
Create aroutdf object indicating p-values, estimates, age-dependent trend
indicators from the regression result
}
