% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_augment}
\alias{tcga_augment}
\title{Final processing step to augment aroutdf with additional columns}
\usage{
tcga_augment(root.dir, annodf, erbnms, erbnms_t1, erbnms_t2, arnms,
  fdr = 0.01, fc = 1.25)
}
\arguments{
\item{root.dir}{root directory of aroutdf object}

\item{annodf}{annotation data}

\item{erbnms}{ER binding gene names}

\item{erbnms_t1}{NKI Tier 1 ER binding genes}

\item{erbnms_t2}{NKI Tier 2 ER binding genes}

\item{arnms}{Age-related gene names}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}
}
\description{
Final processing step to augment aroutdf with additional columns
}
