% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_fet}
\alias{tcga_fet}
\title{Run Fisher's Exact Test on TCGA studies for specified FDR level}
\usage{
tcga_fet(root.dir, fdr = 0.01, fc = 1.25, file.name = NULL,
  suffix = NULL, nki = FALSE)
}
\arguments{
\item{root.dir}{root directory of aroutdf object}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}

\item{file.name}{core file name}

\item{suffix}{suffix for file name}

\item{nki}{logical; if \code{TRUE} use NKI ER Binding gene list specifications}
}
\description{
Run Fisher's Exact Test on TCGA studies for specified FDR level
}
