% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_lm}
\alias{tcga_lm}
\title{TCGA linear regression model scatterplots}
\usage{
tcga_lm(root.dir, file.name, sedf, tcgaarnms, erbnms, age = 60, fdr = 0.01,
  fc = 1.25)
}
\arguments{
\item{root.dir}{root directory of aroutdf object}

\item{file.name}{core file name}

\item{sedf}{analytical data}

\item{tcgaarnms}{TCGA age-related probe names}

\item{erbnms}{ER-binding probe names}

\item{age}{age threshold for calculating biological significance}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}
}
\description{
TCGA linear regression model scatterplots
}
