% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_load}
\alias{tcga_load}
\title{Load TCGA data}
\usage{
tcga_load(organ = c("Br", "Kidney", "Lung", "Prostate", "Thyroid"),
  spec = FALSE, overwrite = FALSE)
}
\arguments{
\item{organ}{organ to obtain TCGA data}

\item{spec}{logical; if \code{TRUE} and objects do not already exist, then column
specification parsing output from \code{readr} is shown.}

\item{overwrite}{logical; only load if set to \code{TRUE} or if the object doesn't
already exist in the current environment. \code{organ}-specific objects \code{cldf}
and \code{ealldf} are always loaded.}
}
\value{
The function is returned invisibly. All data objects are loaded into
the current environment.
}
\description{
Load TCGA data pertaining to a specific organ into the current environment.
}
\details{
The data directories need to have stable, constant paths because the function
hard-codes the directory and file name for each data object. The data objects
are:
\itemize{
\item \code{cldf}: clinical data
\item \code{ealldf}: expression data
\item \code{annodf}: annotation data
\item \code{ardf}: age-related genes
\item \code{ezdf}: EZH2-related genes
\item \code{ebdf}: ER binding genes frmo ChIP-Seq
\item \code{ebdf_t1}: NKI Tier 1 ER binding genes
\item \code{ebdf_t2}: NKI Tier 2 ER binding genes
\item \code{icdf}: IntClust groups (only for breast cancer)
}

Only \code{cldf} and \code{ealldf} are dependent on \code{organ}. An additional \code{icdf} data
object is loaded when \code{organ == "Br"}.
}
\author{
Derek Chiu
}
