% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{tcga_manhattan}
\alias{tcga_manhattan}
\title{Make a manhattan plot specifically for data structures in TCGA studies}
\usage{
tcga_manhattan(data, root.dir, plot.suffix, title)
}
\arguments{
\item{data}{list of data objects for plotting}

\item{root.dir}{root directory to save plots}

\item{plot.suffix}{suffix at end of plot file name}

\item{title}{plot title}
}
\description{
Make a manhattan plot specifically for data structures in TCGA studies
}
