% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_modify}
\alias{tcga_modify}
\title{Modify the biological significance conditions for the aroutdf object}
\usage{
tcga_modify(aroutdf, sedf, root.dir, age = 60, fdr = 0.01, fc = 1.25,
  subset = FALSE, file.name = NULL, suffix = NULL)
}
\arguments{
\item{aroutdf}{age-related output data frame}

\item{sedf}{analytical data}

\item{root.dir}{root directory of aroutdf object}

\item{age}{age threshold for calculating biological significance}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}

\item{subset}{if \code{TRUE}, also save subsetted aroutdf results}

\item{file.name}{core file name}

\item{suffix}{suffix for file name}
}
\description{
Modify the biological significance conditions for the aroutdf object
}
