% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_path}
\alias{tcga_path}
\title{Read in a TCGA object from path specified by four parameters}
\usage{
tcga_path(root.dir, file.name, fdr, fc, sub.dir = NULL, suffix = NULL,
  extension = "csv")
}
\arguments{
\item{root.dir}{root directory of aroutdf object}

\item{file.name}{core file name}

\item{fdr}{numeric vector of false discovery rates to obtain results for}

\item{fc}{numeric vector of fold change values to obtain results for}

\item{sub.dir}{subdirectory}

\item{suffix}{suffix for full name, right before extension}

\item{extension}{file extension}
}
\description{
Read in a TCGA object from path specified by four parameters
}
