% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga_process}
\alias{tcga_process}
\title{Process TCGA data}
\usage{
tcga_process(cldf, ealldf, annodf, ardf, ezdf, ebdf, ebdf_t1, ebdf_t2,
  icdf = NULL, gender = c("all", "male", "female"))
}
\arguments{
\item{cldf}{clinical data}

\item{ealldf}{expression data}

\item{annodf}{annotation data}

\item{ardf}{age related genes}

\item{ezdf}{EZH2 related genes}

\item{ebdf}{ER binding genes from ChIP-Seq}

\item{ebdf_t1}{NKI Tier 1 ER binding genes}

\item{ebdf_t2}{NKI Tier 2 ER binding genes}

\item{icdf}{IntClust groups (leave \code{NULL} unless for breast cancer)}

\item{gender}{Either "all", "male", or "female" subsets}
}
\value{
A list with elements
\itemize{
\item \code{exandf}: expression data restricted to annotation data
\item \code{sedf}: data for regression analysis
\item \code{erbnms}: ER binding gene names
\item \code{arnms}: Age-related gene names
\item \code{tcgaarnms}: TCGA gene names
}
}
\description{
Process TCGA data sources to produce outputs and results ready for analysis.
}
\details{
The data sources loaded from \code{tcga_load} are processed so that we may use the
outputs and results for age-associated biomarker analysis.
}
