% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{tcga_volcano}
\alias{tcga_volcano}
\title{Make a volcano plot specifically for data structures in TCGA studies}
\usage{
tcga_volcano(data, root.dir, plot.suffix, title, nki = FALSE, ...)
}
\arguments{
\item{data}{list of data objects for plotting}

\item{root.dir}{root directory to save plots}

\item{plot.suffix}{suffix at end of plot file name}

\item{title}{plot title}

\item{nki}{if \code{TRUE}, use ER binding gene lists for NKI Tier 1/2 instead of
ChIP-Seq}

\item{...}{named arguments to pass to \code{volcano_plot}}
}
\description{
Make a volcano plot specifically for data structures in TCGA studies
}
