% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_ct}
\alias{tidy_ct}
\title{Tidy contingency tables}
\usage{
tidy_ct(table, workspace = 1e+06)
}
\arguments{
\item{table}{contingency table. Cross tabulation is performed by
\code{\link[descr:CrossTable]{descr::CrossTable()}}.}

\item{workspace}{fisher exact test workspace. Increase as needed.}
}
\description{
Cross tabulation with Chi-Squared Test and Fisher Exact Test results
}
\details{
The Fisher Exact Test is recalculated when there are errors pertaining to an
insufficient workspace ("LDKEY is too small for this problem.")
}
