% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_fet}
\alias{tidy_fet}
\title{Tidy Fisher Exact Test results}
\usage{
tidy_fet(object, digits = 4)
}
\arguments{
\item{object}{object of class "htest" with FET results}

\item{digits}{number of digits to round statistics}
}
\description{
Extract relevant FET results from object of class "htest" into a table.
}
\details{
The estimated odds ratio statistic and corresponding confidence interval only
apply to the 2 by 2 case. The FET p-value is returned for contingency tables
of any dimension.
}
