% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_genes}
\alias{tidy_genes}
\title{Tidy gene names}
\usage{
tidy_genes(data, var)
}
\arguments{
\item{data}{ER binding gene dataset}

\item{var}{variable name with gene symbols in \code{data}}
}
\description{
Extract gene names from ER binding gene lists.
}
\details{
Gene names with the empty string \code{""} or \code{NA} are removed. Duplicate gene
names are removed and finally the gene names are sorted.
}
