% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot}
\usage{
volcano_plot(data, x, y, group, label, subset, legend = NULL,
  color = rainbow(3), title = "", xlab = NULL, ylab = NULL,
  xlim = NULL, ylim = NULL, seed = 1, filename = NULL, style = c("ss",
  "gg"), thin = FALSE)
}
\arguments{
\item{data}{data object with regression fit result variables}

\item{x}{log2 fold change}

\item{y}{adjusted p-value (preferrably in -log10 scale)}

\item{group}{grouping variable of interest (e.g. ER binding group)}

\item{label}{gene variable for labelling points}

\item{subset}{subset of \code{data} used for labelling points}

\item{legend}{character vector for legend key items}

\item{color}{character vector for colour of each legend key item}

\item{title}{title of plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{seed}{reproducible text repel direction}

\item{filename}{file path to save figure to (PDF)}

\item{style}{use \code{\link[graphics:smoothScatter]{graphics::smoothScatter()}} or \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plotting
device?}

\item{thin}{logical; if \code{TRUE}, a clustering is applied on the non
age-dependent points and only certain low/high density points are plotted.
Smoothing with \code{\link[graphics:smoothScatter]{graphics::smoothScatter()}} is not used.}
}
\value{
a ggplot object storing the volcano plot
}
\description{
Volcano Plot
}
\author{
Derek Chiu
}
